+++ src/main/resources/templates/trade/admin/trade_admin.html
@@ -21,8 +21,8 @@
         </nav>
         <h1 class="mb-2">거래 관리자</h1>
     <div class="col">
-        <a class="btn btn-primary btn-lg" href="/trade/trade_admin_application" role="button">상품 등록 신청 목록</a>
-        <a class="btn btn-primary btn-lg" href="/trade/trade_adv_list" role="button">광고 리스트</a>
+        <a class="btn btn-primary btn-lg" href="/admin/trade_admin_application" role="button">상품 등록 신청 목록</a>
+        <a class="btn btn-primary btn-lg" href="/admin/trade_adv_list" role="button">광고 리스트</a>
     </div>
     </div>
 </main>
