+++ src/main/resources/templates/trade/admin/trade_admin_application.html
@@ -25,13 +25,13 @@
         </nav>
         <nav class="navbar navbar-expand-lg bg-body-tertiary">
             <div class="container-fluid">
-                <a class="navbar-brand me-5" href="/trade/trade_admin_application">상품 등록 신청 목록</a>
+                <a class="navbar-brand me-5" href="/admin/trade_admin_application">상품 등록 신청 목록</a>
             </div>
         </nav>
         <div class="container text-center">
             <div class="row row-cols-3 mt-5">
                 <div class="col-md-4 mb-5" th:each="trade : ${mainlist}">
-                    <a th:href="@{/trade/trade_admin_approve(tradeNo=${trade.tradeNo})}" name="tradeNo">
+                    <a th:href="@{/admin/trade_admin_approve(tradeNo=${trade.tradeNo})}" name="tradeNo">
                         <img th:src="@{/images/{image}(image=${trade.storedImageName})}" class="img-thumbnail" style="width:300px; height:300px">
                         <br>
                         <span th:text="${trade.title}">Title</span>
