+++ src/main/resources/templates/trade/admin/trade_adv_detail.html
@@ -41,8 +41,8 @@
     <div align="center">
         <hr class="border border-dark">
         <div class="btn-group" aria-label="Small button group">
-            <a th:href="@{/trade/trade_adv_delete(advNo=${advDetail.advNo})}" class="btn bg-body-secondary">삭제</a>
-            <a href="/trade/trade_adv_list" class="btn bg-body-secondary">목록으로</a>
+            <a th:href="@{/admin/trade_adv_delete(advNo=${advDetail.advNo})}" class="btn bg-body-secondary">삭제</a>
+            <a href="/admin/trade_adv_list" class="btn bg-body-secondary">목록으로</a>
         </div>
     </div>
     </div>
