+++ src/main/resources/templates/trade/admin/trade_adv_regist.html
@@ -21,7 +21,7 @@
         </ol>
     </nav>
     <h1 align="center">거래 광고 등록</h1>
-    <form action="/trade/trade_adv_regist" method="post" enctype="multipart/form-data">
+    <form action="/admin/trade_adv_regist" method="post" enctype="multipart/form-data">
         <div class="input-group">
             <span class="input-group-text" id="inputGroup-sizing-lg" >제목</span>
             <input type="text" class="form-control" aria-label="Sizing example input" aria-describedby="inputGroup-sizing-lg" style="width: 70%;" name="title">
