+++ src/main/resources/templates/trade/trade_main.html
@@ -108,7 +108,7 @@
             </nav>
         </div>
         <th:block sec:authorize="hasRole('ADMIN')">
-            <a class="btn btn-primary btn-lg" href="/trade/trade_admin" role="button">거래 관리자 이동</a>
+            <a class="btn btn-primary btn-lg" href="/admin/trade_admin" role="button">거래 관리자 이동</a>
         </th:block>
 
     </div>
